+++ changelog.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Breaking
 - MoreEvents: rename method `eventId` to `calculateEventId`
 - MoreIdentities: remove methods `fromHex` and `of`
-- build Metadata class from protobuf definition
+- externalize modules `nostr-proto` and `nostr-proto-json`
 
 ## [0.1.0] - 2024-12-21